# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateShipRequest:
    properties:
      beam:
        minimum: 0
        type: number
      build_year:
        maximum: 2100
        minimum: 1800
        type: integer
      builder:
        maxLength: 255
        type: string
      call_sign:
        maxLength: 20
        type: string
      certificate_expiry:
        type: integer
      classification_society:
        maxLength: 255
        type: string
      crew_capacity:
        minimum: 0
        type: integer
      current_latitude:
        maximum: 90
        minimum: -90
        type: number
      current_longitude:
        maximum: 180
        minimum: -180
        type: number
      deadweight_tonnage:
        minimum: 0
        type: number
      draft:
        minimum: 0
        type: number
      flag_state:
        maxLength: 100
        type: string
      gross_tonnage:
        minimum: 0
        type: number
      imo_number:
        maxLength: 20
        type: string
      insurance_expiry:
        type: integer
      last_inspection:
        type: integer
      length:
        minimum: 0
        type: number
      max_speed:
        minimum: 0
        type: number
      mmsi:
        maxLength: 20
        type: string
      net_tonnage:
        minimum: 0
        type: number
      next_inspection:
        type: integer
      notes:
        maxLength: 1000
        type: string
      operator_id:
        type: string
      passenger_capacity:
        minimum: 0
        type: integer
      port_of_registry:
        maxLength: 255
        type: string
      ship_name:
        maxLength: 255
        type: string
      ship_type:
        maxLength: 100
        type: string
    required:
    - flag_state
    - operator_id
    - port_of_registry
    - ship_name
    - ship_type
    type: object
  model.LoginUserRequest:
    properties:
      password:
        maxLength: 100
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - password
    - username
    type: object
  model.PageMetadata:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total:
        type: integer
    type: object
  model.RegisterUserRequest:
    properties:
      email:
        maxLength: 255
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      phone:
        maxLength: 20
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  model.SwaggerPageResponse:
    properties:
      data:
        items: {}
        type: array
      errors:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PageMetadata'
      success:
        type: boolean
    type: object
  model.SwaggerWebResponse:
    properties:
      data: {}
      errors:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/model.PageMetadata'
      success:
        type: boolean
    type: object
...
  model.UpdateShipRequest:
    properties:
      beam:
        minimum: 0
        type: number
      build_year:
        maximum: 2100
        minimum: 1800
        type: integer
      builder:
        maxLength: 255
        type: string
      call_sign:
        maxLength: 20
        type: string
      certificate_expiry:
        type: integer
      classification_society:
        maxLength: 255
        type: string
      crew_capacity:
        minimum: 0
        type: integer
      current_latitude:
        maximum: 90
        minimum: -90
        type: number
      current_longitude:
        maximum: 180
        minimum: -180
        type: number
      deadweight_tonnage:
        minimum: 0
        type: number
      draft:
        minimum: 0
        type: number
      flag_state:
        maxLength: 100
        type: string
      gross_tonnage:
        minimum: 0
        type: number
      imo_number:
        maxLength: 20
        type: string
      insurance_expiry:
        type: integer
      is_active:
        type: boolean
      last_inspection:
        type: integer
      length:
        minimum: 0
        type: number
      max_speed:
        minimum: 0
        type: number
      mmsi:
        maxLength: 20
        type: string
      net_tonnage:
        minimum: 0
        type: number
      next_inspection:
        type: integer
      notes:
        maxLength: 1000
        type: string
      passenger_capacity:
        minimum: 0
        type: integer
      port_of_registry:
        maxLength: 255
        type: string
      ship_name:
        maxLength: 255
        type: string
      ship_type:
        maxLength: 100
        type: string
      status:
        enum:
        - active
        - inactive
        - maintenance
        type: string
    type: object
  model.UpdateUserRequest:
    properties:
      avatar:
        maxLength: 500
        type: string
      email:
        maxLength: 255
        type: string
      first_name:
        maxLength: 100
        type: string
      is_active:
        type: boolean
      last_name:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      phone:
        maxLength: 20
        type: string
      username:
        maxLength: 100
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Golang Clean Architecture.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MKP Boarding Test API
  version: "1.0"
paths:
...
  /api/ships:
    get:
      consumes:
      - application/json
      description: Get list of ships with optional filtering
      parameters:
      - description: Filter by operator ID
        in: query
        name: operator_id
        type: string
      - description: Filter by ship name
        in: query
        name: ship_name
        type: string
      - description: Filter by flag state
        in: query
        name: flag_state
        type: string
      - description: Filter by ship type
        in: query
        name: ship_type
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - default: true
        description: Filter by active status
        in: query
        name: is_active
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of ships
          schema:
            $ref: '#/definitions/model.SwaggerPageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: List ships
      tags:
      - Ships
    post:
      consumes:
      - application/json
      description: Create a new ship with detailed information
      parameters:
      - description: Create ship request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateShipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ship created successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Create a new ship
      tags:
      - Ships
  /api/ships/{shipId}:
    delete:
      consumes:
      - application/json
      description: Delete ship by ship ID
      parameters:
      - description: Ship ID
        in: path
        name: shipId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ship deleted successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Ship not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Delete ship
      tags:
      - Ships
    get:
      consumes:
      - application/json
      description: Get ship details by ship ID
      parameters:
      - description: Ship ID
        in: path
        name: shipId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ship details
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Ship not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Get ship by ID
      tags:
      - Ships
    put:
      consumes:
      - application/json
      description: Update ship information by ship ID
      parameters:
      - description: Ship ID
        in: path
        name: shipId
        required: true
        type: string
      - description: Update ship request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateShipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ship updated successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "404":
          description: Ship not found
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Update ship
      tags:
      - Ships
  /api/users:
    delete:
      consumes:
      - application/json
      description: Logout current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Users
  /api/users/_current:
    get:
      consumes:
      - application/json
      description: Get current authenticated user information
      produces:
      - application/json
      responses:
        "200":
          description: Current user details
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update current authenticated user information
      parameters:
      - description: Update user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - Users
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username/email and password
      parameters:
      - description: Login user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      summary: User login
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, email, and password
      parameters:
      - description: Register user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.SwaggerWebResponse'
      summary: Register a new user
      tags:
      - Auth
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"